:+1: Products ページのパスを products に変更
