+++ apps/frontend/src/app/products/page.tsx
@@ -44,7 +44,7 @@ const Products: NextPage = () => {
     <ContentsLayout
       description=''
       pageType='article'
-      pageUrl={PagePath.tools(true)}
+      pageUrl={PagePath.products(true)}
     >
       <h1 className={clsx('text-3xl')}>Products</h1>
       <div className='mt-8 grid gap-4 grid-cols-[repeat(auto-fit,minmax(400px,_1fr))]'>
