+++ apps/frontend/src/components/Assets/Products.tsx
@@ -1,0 +1,26 @@
+import React, { CSSProperties } from 'react';
+
+interface Props {
+  className?: string;
+  style?: CSSProperties;
+}
+
+export const Products: React.FC<Props> = ({ className, style }) => {
+  return (
+    <svg
+      width='386'
+      height='72'
+      viewBox='0 0 386 72'
+      fill='none'
+      xmlns='http://www.w3.org/2000/svg'
+      className={className}
+      style={style}
+    >
+      <path
+        d='M0.544 71V2.84H28.768C31.712 2.84 34.4 3.448 36.832 4.664C39.328 5.88 41.472 7.54399 43.264 9.65599C45.056 11.704 46.464 14.008 47.488 16.568C48.512 19.064 49.024 21.624 49.024 24.248C49.024 27.96 48.192 31.48 46.528 34.808C44.928 38.072 42.656 40.728 39.712 42.776C36.768 44.824 33.312 45.848 29.344 45.848H7.264V71H0.544ZM7.264 39.896H29.056C31.744 39.896 34.08 39.16 36.064 37.688C38.048 36.216 39.584 34.296 40.672 31.928C41.76 29.56 42.304 27 42.304 24.248C42.304 21.432 41.664 18.84 40.384 16.472C39.104 14.104 37.408 12.248 35.296 10.904C33.248 9.496 30.976 8.792 28.48 8.792H7.264V39.896ZM83.8788 26.744C79.5268 26.872 75.6868 28.088 72.3588 30.392C69.0948 32.696 66.7908 35.864 65.4468 39.896V71H58.9188V20.984H65.0628V32.984C66.7908 29.464 69.0628 26.616 71.8788 24.44C74.7588 22.264 77.7988 21.048 80.9988 20.792C81.6388 20.728 82.1828 20.696 82.6308 20.696C83.1428 20.696 83.5588 20.728 83.8788 20.792V26.744ZM112.293 71.96C108.709 71.96 105.381 71.288 102.309 69.944C99.3013 68.536 96.6773 66.648 94.4373 64.28C92.2613 61.848 90.5653 59.096 89.3493 56.024C88.1333 52.888 87.5253 49.592 87.5253 46.136C87.5253 42.552 88.1333 39.224 89.3493 36.152C90.5653 33.016 92.2933 30.264 94.5333 27.896C96.7733 25.464 99.3973 23.576 102.405 22.232C105.477 20.824 108.805 20.12 112.389 20.12C115.973 20.12 119.269 20.824 122.277 22.232C125.285 23.576 127.909 25.464 130.149 27.896C132.389 30.264 134.117 33.016 135.333 36.152C136.549 39.224 137.157 42.552 137.157 46.136C137.157 49.592 136.549 52.888 135.333 56.024C134.117 59.096 132.389 61.848 130.149 64.28C127.973 66.648 125.349 68.536 122.277 69.944C119.269 71.288 115.941 71.96 112.293 71.96ZM94.1493 46.232C94.1493 49.944 94.9493 53.336 96.5493 56.408C98.2133 59.416 100.421 61.816 103.173 63.608C105.925 65.336 108.965 66.2 112.293 66.2C115.621 66.2 118.661 65.304 121.413 63.512C124.165 61.72 126.373 59.288 128.037 56.216C129.701 53.08 130.533 49.688 130.533 46.04C130.533 42.328 129.701 38.936 128.037 35.864C126.373 32.792 124.165 30.36 121.413 28.568C118.661 26.776 115.621 25.88 112.293 25.88C108.965 25.88 105.925 26.808 103.173 28.664C100.485 30.52 98.3093 32.984 96.6453 36.056C94.9813 39.064 94.1493 42.456 94.1493 46.232ZM144.715 46.136C144.715 41.4 145.675 37.08 147.595 33.176C149.579 29.208 152.299 26.04 155.755 23.672C159.211 21.304 163.147 20.12 167.563 20.12C171.723 20.12 175.435 21.24 178.699 23.48C182.027 25.656 184.619 28.312 186.475 31.448V0.919998H193.003V63.128C193.003 64.024 193.195 64.696 193.579 65.144C193.963 65.528 194.539 65.752 195.307 65.816V71C194.027 71.192 193.035 71.256 192.331 71.192C190.987 71.128 189.803 70.584 188.779 69.56C187.755 68.536 187.243 67.416 187.243 66.2V61.304C185.259 64.568 182.571 67.16 179.179 69.08C175.787 71 172.267 71.96 168.619 71.96C165.163 71.96 161.963 71.256 159.019 69.848C156.139 68.44 153.611 66.52 151.435 64.088C149.323 61.656 147.659 58.904 146.443 55.832C145.291 52.696 144.715 49.464 144.715 46.136ZM186.475 53.816V38.36C185.643 36.056 184.299 33.976 182.443 32.12C180.587 30.2 178.475 28.696 176.107 27.608C173.803 26.456 171.499 25.88 169.195 25.88C166.507 25.88 164.075 26.456 161.899 27.608C159.723 28.696 157.835 30.2 156.235 32.12C154.699 34.04 153.515 36.216 152.683 38.648C151.851 41.08 151.435 43.576 151.435 46.136C151.435 48.824 151.883 51.384 152.779 53.816C153.739 56.248 155.083 58.424 156.811 60.344C158.539 62.2 160.523 63.64 162.763 64.664C165.003 65.688 167.435 66.2 170.059 66.2C171.723 66.2 173.451 65.88 175.243 65.24C177.035 64.6 178.731 63.736 180.331 62.648C181.995 61.496 183.371 60.152 184.459 58.616C185.547 57.08 186.219 55.48 186.475 53.816ZM207.314 50.36V20.984H213.842V49.208C213.842 54.904 214.802 59.16 216.722 61.976C218.706 64.792 221.618 66.2 225.458 66.2C228.018 66.2 230.482 65.656 232.85 64.568C235.282 63.416 237.426 61.816 239.282 59.768C241.138 57.656 242.546 55.224 243.506 52.472V20.984H250.034V63.128C250.034 64.024 250.226 64.696 250.61 65.144C250.994 65.528 251.57 65.752 252.338 65.816V71C251.57 71.064 250.962 71.096 250.514 71.096C250.066 71.16 249.65 71.192 249.266 71.192C247.986 71.064 246.866 70.552 245.906 69.656C245.01 68.76 244.53 67.672 244.466 66.392L244.274 59.672C242.098 63.512 239.122 66.52 235.346 68.696C231.634 70.872 227.602 71.96 223.25 71.96C218.002 71.96 214.034 70.136 211.346 66.488C208.658 62.84 207.314 57.464 207.314 50.36ZM286.49 71.96C282.842 71.96 279.482 71.288 276.41 69.944C273.402 68.536 270.746 66.616 268.442 64.184C266.202 61.752 264.442 58.968 263.162 55.832C261.946 52.696 261.338 49.368 261.338 45.848C261.338 41.112 262.394 36.792 264.506 32.888C266.618 28.984 269.562 25.88 273.338 23.576C277.114 21.272 281.466 20.12 286.394 20.12C291.066 20.12 295.194 21.208 298.778 23.384C302.362 25.496 305.018 28.376 306.746 32.024L300.41 34.04C299.002 31.48 297.018 29.496 294.458 28.088C291.962 26.616 289.178 25.88 286.106 25.88C282.778 25.88 279.738 26.744 276.986 28.472C274.234 30.2 272.026 32.568 270.362 35.576C268.762 38.584 267.962 42.008 267.962 45.848C267.962 49.624 268.794 53.048 270.458 56.12C272.122 59.192 274.33 61.656 277.082 63.512C279.834 65.304 282.874 66.2 286.202 66.2C288.378 66.2 290.458 65.816 292.442 65.048C294.49 64.28 296.25 63.256 297.722 61.976C299.258 60.632 300.282 59.192 300.794 57.656L307.226 59.576C306.33 61.944 304.826 64.088 302.714 66.008C300.666 67.864 298.234 69.336 295.418 70.424C292.666 71.448 289.69 71.96 286.49 71.96ZM341.481 68.6C340.969 68.792 340.169 69.144 339.081 69.656C337.993 70.168 336.681 70.616 335.145 71C333.609 71.384 331.945 71.576 330.153 71.576C328.297 71.576 326.537 71.224 324.873 70.52C323.273 69.816 321.993 68.76 321.033 67.352C320.073 65.88 319.593 64.088 319.593 61.976V26.168H312.681V20.984H319.593V4.088H326.121V20.984H337.641V26.168H326.121V60.152C326.249 61.944 326.889 63.288 328.041 64.184C329.257 65.08 330.633 65.528 332.169 65.528C333.961 65.528 335.593 65.24 337.065 64.664C338.537 64.024 339.433 63.576 339.753 63.32L341.481 68.6ZM366.054 71.96C361.958 71.96 358.15 71.288 354.63 69.944C351.11 68.6 348.07 66.552 345.51 63.8L348.198 59.192C351.014 61.816 353.83 63.736 356.646 64.952C359.526 66.104 362.566 66.68 365.766 66.68C369.67 66.68 372.838 65.912 375.27 64.376C377.702 62.776 378.918 60.504 378.918 57.56C378.918 55.576 378.31 54.072 377.094 53.048C375.942 51.96 374.246 51.096 372.006 50.456C369.83 49.752 367.206 49.016 364.134 48.248C360.678 47.288 357.766 46.296 355.398 45.272C353.094 44.184 351.334 42.84 350.118 41.24C348.966 39.576 348.39 37.432 348.39 34.808C348.39 31.544 349.19 28.856 350.79 26.744C352.454 24.568 354.694 22.936 357.51 21.848C360.39 20.696 363.59 20.12 367.109 20.12C370.95 20.12 374.342 20.728 377.286 21.944C380.23 23.16 382.63 24.856 384.486 27.032L381.318 31.448C379.526 29.4 377.382 27.896 374.886 26.936C372.454 25.912 369.734 25.4 366.726 25.4C364.678 25.4 362.726 25.688 360.87 26.264C359.014 26.776 357.478 27.672 356.262 28.952C355.11 30.168 354.534 31.864 354.534 34.04C354.534 35.832 354.982 37.24 355.878 38.264C356.774 39.224 358.118 40.056 359.91 40.76C361.702 41.4 363.91 42.104 366.534 42.872C370.31 43.896 373.606 44.952 376.422 46.04C379.238 47.064 381.414 48.408 382.95 50.072C384.486 51.736 385.254 54.072 385.254 57.08C385.254 61.752 383.494 65.4 379.974 68.024C376.454 70.648 371.814 71.96 366.054 71.96Z'
+        fill='none'
+        className='fill-current'
+      />
+    </svg>
+  );
+};
