--- apps/frontend/src/components/Assets/Tools.tsx
@@ -1,26 +1,0 @@
-import { CSSProperties } from 'react';
-
-interface Props {
-  className?: string;
-  style?: CSSProperties;
-}
-
-export const Tools = (props: Props): JSX.Element => {
-  return (
-    <svg
-      width='230'
-      height='72'
-      viewBox='0 0 230 72'
-      fill='none'
-      xmlns='http://www.w3.org/2000/svg'
-      className={props.className}
-      style={props.style}
-    >
-      <path
-        d='M55.928 8.792H31.64V71H24.92V8.792H0.632V2.84H55.928V8.792ZM75.3245 71.96C71.7405 71.96 68.4125 71.288 65.3405 69.944C62.3325 68.536 59.7085 66.648 57.4685 64.28C55.2925 61.848 53.5965 59.096 52.3805 56.024C51.1645 52.888 50.5565 49.592 50.5565 46.136C50.5565 42.552 51.1645 39.224 52.3805 36.152C53.5965 33.016 55.3245 30.264 57.5645 27.896C59.8045 25.464 62.4285 23.576 65.4365 22.232C68.5085 20.824 71.8365 20.12 75.4205 20.12C79.0045 20.12 82.3005 20.824 85.3085 22.232C88.3165 23.576 90.9405 25.464 93.1805 27.896C95.4205 30.264 97.1485 33.016 98.3645 36.152C99.5805 39.224 100.189 42.552 100.189 46.136C100.189 49.592 99.5805 52.888 98.3645 56.024C97.1485 59.096 95.4205 61.848 93.1805 64.28C91.0045 66.648 88.3805 68.536 85.3085 69.944C82.3005 71.288 78.9725 71.96 75.3245 71.96ZM57.1805 46.232C57.1805 49.944 57.9805 53.336 59.5805 56.408C61.2445 59.416 63.4525 61.816 66.2045 63.608C68.9565 65.336 71.9965 66.2 75.3245 66.2C78.6525 66.2 81.6925 65.304 84.4445 63.512C87.1965 61.72 89.4045 59.288 91.0685 56.216C92.7325 53.08 93.5645 49.688 93.5645 46.04C93.5645 42.328 92.7325 38.936 91.0685 35.864C89.4045 32.792 87.1965 30.36 84.4445 28.568C81.6925 26.776 78.6525 25.88 75.3245 25.88C71.9965 25.88 68.9565 26.808 66.2045 28.664C63.5165 30.52 61.3405 32.984 59.6765 36.056C58.0125 39.064 57.1805 42.456 57.1805 46.232ZM132.418 71.96C128.834 71.96 125.506 71.288 122.434 69.944C119.426 68.536 116.802 66.648 114.562 64.28C112.386 61.848 110.69 59.096 109.474 56.024C108.258 52.888 107.65 49.592 107.65 46.136C107.65 42.552 108.258 39.224 109.474 36.152C110.69 33.016 112.418 30.264 114.658 27.896C116.898 25.464 119.522 23.576 122.53 22.232C125.602 20.824 128.93 20.12 132.514 20.12C136.098 20.12 139.394 20.824 142.402 22.232C145.41 23.576 148.034 25.464 150.274 27.896C152.514 30.264 154.242 33.016 155.458 36.152C156.674 39.224 157.282 42.552 157.282 46.136C157.282 49.592 156.674 52.888 155.458 56.024C154.242 59.096 152.514 61.848 150.274 64.28C148.098 66.648 145.474 68.536 142.402 69.944C139.394 71.288 136.066 71.96 132.418 71.96ZM114.274 46.232C114.274 49.944 115.074 53.336 116.674 56.408C118.338 59.416 120.546 61.816 123.298 63.608C126.05 65.336 129.09 66.2 132.418 66.2C135.746 66.2 138.786 65.304 141.538 63.512C144.29 61.72 146.498 59.288 148.162 56.216C149.826 53.08 150.658 49.688 150.658 46.04C150.658 42.328 149.826 38.936 148.162 35.864C146.498 32.792 144.29 30.36 141.538 28.568C138.786 26.776 135.746 25.88 132.418 25.88C129.09 25.88 126.05 26.808 123.298 28.664C120.61 30.52 118.434 32.984 116.77 36.056C115.106 39.064 114.274 42.456 114.274 46.232ZM168.488 0.919998H175.016V59.672C175.016 62.168 175.464 63.8 176.36 64.568C177.256 65.336 178.344 65.72 179.624 65.72C180.776 65.72 181.928 65.592 183.08 65.336C184.232 65.08 185.224 64.824 186.056 64.568L187.208 69.848C185.992 70.36 184.488 70.776 182.696 71.096C180.968 71.416 179.464 71.576 178.184 71.576C175.24 71.576 172.872 70.712 171.08 68.984C169.352 67.256 168.488 64.888 168.488 61.88V0.919998ZM210.772 71.96C206.676 71.96 202.868 71.288 199.348 69.944C195.828 68.6 192.788 66.552 190.228 63.8L192.916 59.192C195.732 61.816 198.548 63.736 201.364 64.952C204.244 66.104 207.284 66.68 210.484 66.68C214.388 66.68 217.556 65.912 219.988 64.376C222.42 62.776 223.636 60.504 223.636 57.56C223.636 55.576 223.028 54.072 221.812 53.048C220.66 51.96 218.964 51.096 216.724 50.456C214.548 49.752 211.924 49.016 208.852 48.248C205.396 47.288 202.484 46.296 200.116 45.272C197.812 44.184 196.052 42.84 194.836 41.24C193.684 39.576 193.108 37.432 193.108 34.808C193.108 31.544 193.908 28.856 195.508 26.744C197.172 24.568 199.412 22.936 202.228 21.848C205.108 20.696 208.308 20.12 211.828 20.12C215.668 20.12 219.06 20.728 222.004 21.944C224.948 23.16 227.348 24.856 229.204 27.032L226.036 31.448C224.244 29.4 222.1 27.896 219.604 26.936C217.172 25.912 214.452 25.4 211.444 25.4C209.396 25.4 207.444 25.688 205.588 26.264C203.732 26.776 202.196 27.672 200.98 28.952C199.828 30.168 199.252 31.864 199.252 34.04C199.252 35.832 199.7 37.24 200.596 38.264C201.492 39.224 202.836 40.056 204.628 40.76C206.42 41.4 208.628 42.104 211.252 42.872C215.028 43.896 218.324 44.952 221.14 46.04C223.956 47.064 226.132 48.408 227.668 50.072C229.204 51.736 229.972 54.072 229.972 57.08C229.972 61.752 228.212 65.4 224.692 68.024C221.172 70.648 216.532 71.96 210.772 71.96Z'
-        fill='none'
-        className='fill-current'
-      />
-    </svg>
-  );
-};
