+++ apps/frontend/src/components/Assets/index.ts
@@ -3,4 +3,4 @@ export * from './Blog';
 export * from './Contact';
 export * from './Logo';
 export * from './ProfileImage';
-export * from './Tools';
+export * from './Products';
